requirements.txt
Flask==2.3.2
JPype1==1.4.1
requests==2.31.0
